Integrating maze with pathfinder

+++ PathFinding/include/Cell.h
@@ -6,7 +6,7 @@ namespace pathfinding {
 class Cell {
    public:
     Cell(){};
-    Cell(const uint32_t& row, const uint32_t& col) : row_{row}, col_{col} {};
+    Cell(const int32_t& row, const int32_t& col) : row_{row}, col_{col} {};
 
     // Compare operator for Cell
     // Essential for this be considered as a key in
@@ -23,12 +23,12 @@ class Cell {
         return (this->row_ != c.row_ || this->col_ != c.col_);
     }
 
-    const uint32_t getRow() const { return row_; }
-    const uint32_t getCol() const { return col_; }
+    const int32_t getRow() const { return row_; }
+    const int64_t getCol() const { return col_; }
 
    private:
-    uint32_t row_;
-    uint32_t col_;
+    int32_t row_;
+    int32_t col_;
 };
 
 // Major Gotcha! Not using inline causes multiple definitions of <<
